backend/requirements.txt
fastapi~=0.115.0
uvicorn[standard]
PyMySQL
sqlalchemy~=2.0.35
cryptography
pydantic~=2.9.2
passlib~=1.7.4
python-jose
bcrypt
python-multipart